+++ src/main/java/com/benmu/framework/adapter/DefaultImageAdapter.java
@@ -87,6 +87,20 @@ public class DefaultImageAdapter {
         ((Activity) context).startActivityForResult(intent, Constant.ImageConstants.IMAGE_PICKER);
     }
 
+    public void cameraUploade(final Context context, UploadImageBean bean) {
+        imagePicker.setImageLoader(new GlideImageLoader());   //设置图片加载器
+        imagePicker.setCrop(true);//允许裁剪
+        imagePicker.setSaveRectangle(true); //是否按矩形区域保存
+        imagePicker.setStyle(CropImageView.Style.RECTANGLE);  //裁剪框的形状
+        imagePicker.setFocusWidth(Constant.ImageConstants.BIGGESTWIDTH);   //裁剪框的宽度。单位像素（圆形自动取宽高最小值）
+        imagePicker.setFocusHeight(Constant.ImageConstants.BIGGESTWIDTH);  //裁剪框的高度。单位像素（圆形自动取宽高最小值）
+        PersistentManager persistentManager = ManagerFactory.getManagerService(PersistentManager
+                .class);
+        persistentManager.setCacheData(Constant.ImageConstants.UPLOAD_IMAGE_BEAN, bean);
+        if (context instanceof Activity) {
+            imagePicker.takePicture((Activity) context, 1001);
+        }
+    }
 
     public void UpMultipleImageData(Context context, ArrayList<ImageItem> items, UploadImageBean
             bean) {