+++ src/main/java/com/benmu/framework/event/DispatchEventCenter.java
@@ -25,6 +25,7 @@ import com.benmu.framework.event.router.EventFinish;
 import com.benmu.framework.event.router.EventGetBackParams;
 import com.benmu.framework.event.router.EventGetParams;
 import com.benmu.framework.event.router.EventOpen;
+import com.benmu.framework.event.router.EventOpenBrowser;
 import com.benmu.framework.event.router.EventRefresh;
 import com.benmu.framework.event.router.EventToMap;
 import com.benmu.framework.event.router.EventWebView;
@@ -158,6 +159,9 @@ public class DispatchEventCenter {
             case WXConstant.WXEventCenter.EVENT_CAMERA_UPLOADIMAGE:
                 new EventCamera().uploadImage(params, context, weexEventBean.getJscallback());
                 break;
+            case WXConstant.WXEventCenter.EVENT_CAMERA_UPLOAD:
+                new EventCamera().cameraUpload(params, context, weexEventBean.getJscallback());
+                break;
 
             case WXConstant.WXEventCenter.EVENT_CAMERA:
                 JSCallback jscallback = weexEventBean.getJscallback();
@@ -206,7 +210,16 @@ public class DispatchEventCenter {
                 new EventGeo().getLocation(context, weexEventBean.getJscallback());
                 break;
             case WXConstant.WXEventCenter.EVENT_WECHATLOGIN:
-                new EventAuth().wechat(context,params,weexEventBean.getJscallback());
+                new EventAuth().wechat(context, params, weexEventBean.getJscallback());
+                break;
+            case WXConstant.WXEventCenter.EVENT_OPENBROWSER:
+                new EventOpenBrowser().openBrowser(context, params, weexEventBean.getJscallback());
+                break;
+            case WXConstant.WXEventCenter.EVENT_COMMUNICATION_SMS:
+                new EventCommunication().sms(weexEventBean.getExpand(), params, context);
+                break;
+            case WXConstant.WXEventCenter.EVENT_COMMUNICATION_CONTACTS:
+                new EventCommunication().contacts(context,weexEventBean.getJscallback());
                 break;
         }
     }