+++ src/main/java/com/benmu/framework/event/shorage/EventGetData.java
@@ -25,9 +25,7 @@ public class EventGetData {
         } else {
             bean.resCode = 0;
         }
-        StorageRaesultBean.Result data = new StorageRaesultBean.Result();
-        data.value = result;
-        bean.data = data;
+        bean.value = result;
         if (jscallback != null) {
             jscallback.invoke(bean);
         }
@@ -43,9 +41,7 @@ public class EventGetData {
         } else {
             bean.resCode = 0;
         }
-        StorageRaesultBean.Result data = new StorageRaesultBean.Result();
-        data.value = result;
-        bean.data = data;
+        bean.value = result;
 
         return bean;
     }