+++ src/main/java/com/benmu/framework/extend/module/RouterModule.java
@@ -49,6 +49,7 @@ public class RouterModule extends WXModule {
         weexEventBean.setJscallback(callback);
         ManagerFactory.getManagerService(DispatchEventManager.class).getBus().post(weexEventBean);
     }
+
     @JSMethod(uiThread = true)
     public void finish(String params, JSCallback callback) {
         WeexEventBean weexEventBean = new WeexEventBean();
@@ -97,15 +98,14 @@ public class RouterModule extends WXModule {
         ManagerFactory.getManagerService(DispatchEventManager.class).getBus().post(weexEventBean);
     }
 
-    @JSMethod
-    public void callPhone(String params) {
+
+    @JSMethod(uiThread = true)
+    public void openBrowser(String params, JSCallback callback) {
         WeexEventBean weexEventBean = new WeexEventBean();
-        weexEventBean.setKey(WXConstant.WXEventCenter.EVENT_CALL);
+        weexEventBean.setKey(WXConstant.WXEventCenter.EVENT_OPENBROWSER);
         weexEventBean.setContext(mWXSDKInstance.getContext());
         weexEventBean.setJsParams(params);
+        weexEventBean.setJscallback(callback);
         ManagerFactory.getManagerService(DispatchEventManager.class).getBus().post(weexEventBean);
-
     }
-
-
 }